passare al fisso per un test

+++ include/cluster.h
@@ -11,12 +11,12 @@ typedef struct cluster_point
     double x;
     double y;
     int time;
-    uint32_t cluster_id;
+    unsigned int cluster_id;
 } cluster_point;
 
-std::vector<point> neighborhood (key_set sigma, const point * center, uint8_t delta);
-// std::vector<std::vector<point>> cluster (key_set sigma, uint8_t mu, uint8_t delta);
-void next_period(std::vector<cluster_point> &temp, int time, key_set sigma, uint8_t mu, float precision, uint8_t delta);
+std::vector<point> neighborhood (key_set sigma, const point * center, unsigned int delta);
+// std::vector<std::vector<point>> cluster (key_set sigma, unsigned int mu, unsigned int delta);
+void next_period(std::vector<cluster_point> &temp, int time, key_set sigma, unsigned int mu, float precision, unsigned int delta);
 void calculate_results(std::vector<cluster_point>& results, key_set& significant_tiles, int delta, int mu, float precision, int current_time);
 
 #endif
